discussion-forum/classessesment-backend-2/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
cachetools==3.1.0
celery==3.1.25
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
Django==2.1.7
django-bootstrap==0.2.4
django-celery==3.2.2
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-taggit==0.24.0
djangorestframework==3.9.2
djongo==1.2.31
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.1
idna==2.8
kombu==3.0.37
oauthlib==3.0.1
ParallelDots==3.2.11
Pillow==5.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pymongo==3.7.2
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
